+++ src/StkTable/types/index.ts
@@ -233,3 +233,7 @@ export type AutoRowHeightConfig<DT> = {
     /** Estimated row height */
     expectedHeight?: number | ((row: DT, index: number) => number);
 };
+
+export type ColResizableConfig<DT extends Record<string, any>> = {
+    disabled: (col: StkTableColumn<DT>) => boolean;
+};